// web/src/api/model/locationResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * nimap API
 * OpenAPI spec version: dev
 */
import type { Address } from './address.ts';
import type { LocationResponseStatus } from './locationResponseStatus.ts';
import type { MultiPolygon } from './multiPolygon.ts';
import type { Point } from './point.ts';

export interface LocationResponse {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  address?: Address;
  area?: MultiPolygon;
  createdAt: string;
  id: string;
  name: string;
  notes?: string;
  parentId?: string;
  representativePoint?: Point;
  representativePointManual: boolean;
  status: LocationResponseStatus;
  updatedAt: string;
}